requirements.txt
anyio==4.4.0
APScheduler==3.10.4
certifi==2024.8.30
exceptiongroup==1.2.2
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.9
pytz==2024.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
typing_extensions==4.12.2
tzlocal==5.2
python-dotenv~=1.0.1
tenacity~=9.0.0